+++ Exercicios-Lista/ex034.py
@@ -1,5 +1,5 @@
 # ler o salario e calcular o aumento. superior de 1250 aumento de 10%, inferiores aumento de 15%.
-sal = int(input('Digite seu salario: R$'))
+sal = float(input('Digite seu salario: R$'))
 if sal >= 1250:
     nsal = sal + (sal* 0.10 )
     print(f'O seu novo salario com o aumento de 10% será de R${nsal}')
